# docs/swagger.yaml
definitions:
  controller.CommonError:
    properties:
      error:
        type: string
    type: object
  controller.CommonResponse:
    properties:
      message:
        type: string
    type: object
  controller.CreateUsersRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      name:
        type: string
    type: object
  models.User:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact:
    email: test@test.com
    name: OnlineShop API Service
    url: http://test.com
  title: OnlineShop API
paths:
  /Users:
    get:
      description: Получение списка всех пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Получение пользователей
      tags:
      - User
  /ping:
    get:
      description: Проверка сервиса
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
      summary: Health-check
      tags:
      - Ping
  /users:
    post:
      description: Добавление нового пользователя
      parameters:
      - description: информация о новом пользователе
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.CreateUsersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Добавление нового пользователя
      tags:
      - Users
  /users/{id}:
    delete:
      description: Удаление пользователя по ID
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Удалить пользователя по ID
      tags:
      - User
    get:
      description: Получение пользователя по ID
      parameters:
      - description: id Пользлвател
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Получить пользователя по ID
      tags:
      - User
    put:
      description: Обновление пользователя по ID
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: информация о пользователе
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.CreateUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Обновить пользователя по ID
      tags:
      - User
swagger: "2.0"